% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly.R
\name{assembly}
\alias{assembly}
\title{Run assembly for fastq data}
\usage{
assembly(outdir, config_file = NULL, fq_table = NULL, min_contig_len = 500)
}
\arguments{
\item{min_contig_len}{}
}
\description{
Run assembly for fastq data
}
