% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{pre.process}
\alias{pre.process}
\title{Run preprocess for raw data}
\usage{
pre.process(raw.data.file, host = "", outdir = "", config_file = NULL)
}
\arguments{
\item{outdir}{}
}
\description{
Run preprocess for raw data
}
